# apps/server/swagger-doc.example.yml
openapi: 3.0.4
info:
  title: OpenSource Together API
  description: |
    Documentation de l'API pour la plateforme OpenSource Together.

    ## Authentification

    L'API utilise SuperTokens avec l'authentification GitHub OAuth.


    TEST TEST TEST
    ### Pour les développeurs Frontend

    **Processus de connexion :**
    1. Rediriger vers `/auth/authorize/github` pour initier la connexion
    2. L'utilisateur est redirigé vers GitHub pour autoriser l'application
    3. GitHub redirige vers `/auth/callback/github` avec le code d'autorisation
    4. Les cookies de session sont automatiquement définis
    5. Utiliser `/auth/session` pour récupérer les infos utilisateur (équivalent `/me`)

    **Cookies de session :**
    - `sAccessToken` : Token d'accès principal
    - `sRefreshToken` : Token de rafraîchissement
    - Ces cookies sont automatiquement gérés par le navigateur

    **Routes protégées :**
    Les routes nécessitant une authentification sont marquées avec le symbole 🔒.

    **Base URL :** `http://localhost:4000`
  version: 1.0.0
servers:
  - url: http://localhost:4000
...
paths:
...
  /projects/{id}:
    get:
      tags:
        - Projects
      summary: Récupérer un projet par ID
      operationId: getProjectById
      parameters:
        - name: id
          in: path
          description: ID du projet à récupérer
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Détails du projet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDto'
        '404':
          description: Projet non trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Project not found'
                  statusCode:
                    type: integer
                    example: 404
    patch:
      tags:
        - Projects
      summary: Mettre à jour un projet existant
      operationId: updateProject
      security:
        - supertokens_session: []
      parameters:
        - name: id
          in: path
          description: ID du projet à mettre à jour
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectDtoRequest'
      responses:
        '200':
          description: Projet mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDto'
        '400':
          description: Erreur lors de la mise à jour du projet
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Message d'erreur spécifique (validation, autorisation, etc.)
                    oneOf:
                      - example: 'Le titre du projet est requis'
                      - example: 'La description du projet doit comporter au minimum 10 caracteres'
                      - example: 'Project not found'
                  statusCode:
                    type: integer
                    example: 400
        '401':
          description: Non authentifié - session requise
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'unauthorised'
                  statusCode:
                    type: integer
                    example: 401
        '404':
          description: Projet non trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Project not found'
                  statusCode:
                    type: integer
                    example: 404
    delete:
      tags:
        - Projects
      summary: 🗑️ Supprimer un projet par ID
      operationId: deleteProject
      description: |
        Supprime définitivement un projet de la plateforme.

        **🔒 Restriction de sécurité :**
        - Seul le propriétaire (créateur) du projet peut le supprimer
        - Vérification automatique de propriété côté serveur

        **⚠️ Action irréversible :**
        Cette action supprime définitivement le projet et toutes ses données associées.
      security:
        - supertokens_session: []
      parameters:
        - name: id
          in: path
          description: ID unique du projet à supprimer
          required: true
          schema:
            type: string
            example: '123e4567-e89b-12d3-a456-426614174000'
      responses:
        '200':
          description: Projet supprimé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Project deleted successfully'
        '401':
          description: Non authentifié - session requise
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'unauthorised'
                  statusCode:
                    type: integer
                    example: 401
        '403':
          description: Accès interdit - vous n'êtes pas le propriétaire de ce projet
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'You are not allowed to delete this project'
                  statusCode:
                    type: integer
                    example: 403
        '404':
          description: Projet non trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Project not found'
                  statusCode:
                    type: integer
                    example: 404
        '400':
          description: Erreur lors de la suppression du projet
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Message d'erreur technique lors de la suppression
                    example: 'Unknown error during project deletion: Database constraint violation'
                  statusCode:
                    type: integer
                    example: 400
  /projects/{id}/roles:
    post:
      tags:
        - Projects
      summary: Créer un nouveau rôle de projet
      operationId: createProjectRole
      description: |
        Crée un nouveau rôle pour un projet existant.
        Nécessite d'être le propriétaire du projet.
      security:
        - supertokens_session: []
      parameters:
        - name: id
          in: path
          description: ID du projet dans lequel créer le rôle
          required: true
          schema:
            type: string
            example: '123e4567-e89b-12d3-a456-426614174000'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRoleDto'
            example:
              roleTitle: 'Frontend Developer'
              description: 'Développeur React expérimenté'
              isFilled: false
              skillSet:
                - id: 'react'
                - id: 'typescript'
      responses:
        '200':
          description: Rôle de projet créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Project role created successfully'
                  role:
                    $ref: '#/components/schemas/ProjectRoleResponseDto'
        '400':
          description: Erreur lors de la création du rôle
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Message d'erreur spécifique
                    oneOf:
                      - example: 'Le titre du rôle est requis'
                      - example: 'Project role already exists'
                      - example: 'Related record not found (e.g., invalid skill)'
                      - example: 'One or more techStacks do not exist: tech1, tech2'
                  statusCode:
                    type: integer
                    example: 400
        '401':
          description: Non authentifié - session requise
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'unauthorised'
                  statusCode:
                    type: integer
                    example: 401
        '403':
          description: Accès interdit - vous n'êtes pas le propriétaire de ce projet
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'You are not allowed to add roles to this project'
                  statusCode:
                    type: integer
                    example: 403
        '404':
          description: Projet non trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Project not found'
                  statusCode:
                    type: integer
                    example: 404
  /projects/{id}/roles/{roleId}:
    patch:
      tags:
        - Projects
      summary: Mettre à jour un rôle de projet existant
      operationId: updateProjectRole
      description: |
        Met à jour les informations d'un rôle spécifique dans un projet.

        **🔒 Sécurité :**
        - Seul le propriétaire du projet peut modifier les rôles
        - Vérification automatique de propriété côté serveur

        **✏️ Modifications possibles :**
        - Titre du rôle
        - Description
        - Statut (pourvu ou non)
        - Compétences techniques requises
      security:
        - supertokens_session: []
      parameters:
        - name: id
          in: path
          description: ID du projet contenant le rôle
          required: true
          schema:
            type: string
            example: '123e4567-e89b-12d3-a456-426614174000'
        - name: roleId
          in: path
          description: ID du rôle à mettre à jour
          required: true
          schema:
            type: string
            example: '789e4567-e89b-12d3-a456-426614174111'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRoleDto'
            example:
              roleTitle: 'Senior Frontend Developer'
              description: 'Développeur React expérimenté avec expertise en TypeScript'
              isFilled: true
              skillSet:
                - id: 'react'
                - id: 'typescript'
      responses:
        '200':
          description: Rôle de projet mis à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Project role updated successfully'
        '400':
          description: Erreur lors de la mise à jour du rôle
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Message d'erreur spécifique (validation, autorisation, etc.)
                    oneOf:
                      - example: 'Le titre du rôle est requis'
                      - example: 'La description du rôle doit comporter au minimum 10 caractères'
                  statusCode:
                    type: integer
                    example: 400
        '401':
          description: Non authentifié - session requise
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'unauthorised'
                  statusCode:
                    type: integer
                    example: 401
        '403':
          description: Accès interdit - vous n'êtes pas le propriétaire de ce projet
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'You are not allowed to update this project role'
                  statusCode:
                    type: integer
                    example: 403
        '404':
          description: Projet ou rôle non trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    oneOf:
                      - example: 'Project not found'
                      - example: 'Role not found in this project'
                  statusCode:
                    type: integer
                    example: 404
    delete:
      tags:
        - Projects
      summary: Supprimer un rôle de projet
      operationId: deleteProjectRole
      description: |
        Supprime un rôle spécifique d'un projet.

        **🔒 Sécurité :**
        - Seul le propriétaire du projet peut supprimer des rôles
        - Vérification automatique de propriété côté serveur

        **⚠️ Action irréversible :**
        Cette action supprime définitivement le rôle et ses associations.
      security:
        - supertokens_session: []
      parameters:
        - name: id
          in: path
          description: ID du projet contenant le rôle
          required: true
          schema:
            type: string
            example: '123e4567-e89b-12d3-a456-426614174000'
        - name: roleId
          in: path
          description: ID du rôle à supprimer
          required: true
          schema:
            type: string
            example: '789e4567-e89b-12d3-a456-426614174111'
      responses:
        '200':
          description: Rôle de projet supprimé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Project role deleted successfully'
        '400':
          description: Erreur lors de la suppression du rôle
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Message d'erreur technique lors de la suppression
                    oneOf:
                      - example: 'Unknown error during role deletion'
                      - example: 'Cannot delete role: it may still be referenced elsewhere'
                  statusCode:
                    type: integer
                    example: 400
        '401':
          description: Non authentifié - session requise
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'unauthorised'
                  statusCode:
                    type: integer
                    example: 401
        '403':
          description: Accès interdit - vous n'êtes pas le propriétaire de ce projet
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'You are not allowed to delete roles from this project'
                  statusCode:
                    type: integer
                    example: 403
        '404':
          description: Projet ou rôle non trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    oneOf:
                      - example: 'Project not found'
                      - example: 'Role not found in this project'
                  statusCode:
                    type: integer
                    example: 404
...
components:
  schemas:
    ProjectRoleResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Identifiant unique du rôle
          example: '789e4567-e89b-12d3-a456-426614174111'
        projectId:
          type: string
          description: Identifiant du projet auquel appartient le rôle
          example: '123e4567-e89b-12d3-a456-426614174000'
        roleTitle:
          type: string
          description: Titre du rôle
          example: 'Frontend Developer'
        description:
          type: string
          description: Description détaillée du rôle
          example: 'Développeur React expérimenté avec expertise en TypeScript'
        isFilled:
          type: boolean
          description: Indique si le rôle est pourvu
          example: false
        skillSet:
          type: array
          items:
            $ref: '#/components/schemas/TechStackDto'
          description: Compétences techniques requises pour ce rôle
    CreateProjectRoleDto:
      type: object
      properties:
        roleTitle:
          type: string
          description: Titre du rôle
          example: 'Frontend Developer'
        description:
          type: string
          description: Description du rôle
          example: 'Développeur React expérimenté'
        isFilled:
          type: boolean
          description: Indique si le rôle est déjà pourvu
          example: false
        skillSet:
          type: array
          items:
            $ref: '#/components/schemas/TechStackDto'
          description: Compétences techniques requises
      required:
        - roleTitle
        - description
        - isFilled
        - skillSet
    UpdateProjectRoleDto:
      type: object
      properties:
        roleTitle:
          type: string
          description: Nouveau titre du rôle
          example: 'Senior Frontend Developer'
        description:
          type: string
          description: Nouvelle description du rôle
          example: 'Développeur React expérimenté avec expertise en TypeScript'
        isFilled:
          type: boolean
          description: État du rôle (pourvu ou non)
          example: true
        skillSet:
          type: array
          items:
            $ref: '#/components/schemas/TechStackDto'
          description: Compétences techniques requises pour ce rôle
    TechStackDto:
      type: object
      properties:
        id:
          type: string
          description: ID de la stack technique
          example: 'typescript'
        name:
          type: string
          description: Nom de la stack technique
          example: 'TypeScript'
        iconUrl:
          type: string
          description: URL de l'icône de la stack technique
          example: 'https://cdn.worldvectorlogo.com/logos/typescript.svg'
      required:
        - id
        - name
        - iconUrl
    ProjectResponseDto:
      type: object
      properties:
        id:
          type: string
          description: ID unique du projet
          example: '123e4567-e89b-12d3-a456-426614174000'
        title:
          type: string
          description: Titre du projet
          example: 'Plateforme E-commerce React'
        description:
          type: string
          description: Description détaillée du projet
          example: 'Application e-commerce moderne avec React et Node.js'
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
          description: Niveau de difficulté du projet
          example: 'MEDIUM'
        link:
          type: string
          description: Lien vers le site web ou démo du projet
          nullable: true
          example: 'https://my-ecommerce-demo.com'
        githubLink:
          type: string
          description: Lien vers le dépôt GitHub du projet
          nullable: true
          example: 'https://github.com/user/ecommerce-platform'
        ownerId:
          type: string
          description: ID de l'utilisateur propriétaire du projet
          example: 'github_user123'
        createdAt:
          type: string
          format: date-time
          description: Date de création du projet
          example: '2024-01-15T10:30:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: Date de dernière modification
          example: '2024-01-20T14:45:00.000Z'
        techStacks:
          type: array
          items:
            $ref: '#/components/schemas/TechStackDto'
          description: Technologies utilisées dans le projet
        projectRoles:
          type: array
          items:
            $ref: '#/components/schemas/ProjectRoleDto'
          description: Rôles recherchés pour le projet
        projectMembers:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMemberDto'
          description: Membres actuels du projet
        owner:
          $ref: '#/components/schemas/GitHubUser'
          description: Informations du propriétaire du projet
      required:
        - id
        - title
        - description
        - difficulty
        - ownerId
        - createdAt
        - updatedAt
        - techStacks
        - projectRoles
        - projectMembers
...
    UpdateProjectDtoRequest:
      type: object
      properties:
        title:
          type: string
          description: Nouveau titre du projet
          example: 'Mon Super Projet Open Source (Mis à jour)'
        description:
          type: string
          description: Nouvelle description du projet
          example: 'Description mise à jour du projet open source.'
        link:
          type: string
          description: Nouveau lien vers le dépôt ou le site du projet
          nullable: true
          example: 'https://github.com/user/my-updated-project'
        status:
          type: string
          description: Nouveau statut du projet
          example: 'Completed'
        techStacks:
          type: array
          items:
            $ref: '#/components/schemas/TechStackDto'
          description: Liste mise à jour des stacks techniques utilisées dans le projet
      required:
        - title
        - description
        - status
        - techStacks
...
    GitHubUser:
      type: object
      properties:
        id:
          type: string
          description: ID utilisateur unique
          example: 'github_123456789'
        email:
          type: string
          description: Email de l'utilisateur GitHub
          example: 'developer@example.com'
        github:
          type: object
          properties:
            login:
              type: string
              description: Nom d'utilisateur GitHub
              example: 'johndoe'
            name:
              type: string
              description: Nom complet de l'utilisateur
              example: 'John Doe'
            avatar_url:
              type: string
              description: URL de l'avatar GitHub
              example: 'https://avatars.githubusercontent.com/u/123456789'
            html_url:
              type: string
              description: URL du profil GitHub public
              example: 'https://github.com/johndoe'
            bio:
              type: string
              description: Biographie GitHub de l'utilisateur
              example: 'Full-stack developer passionate about open source'
            public_repos:
              type: integer
              description: Nombre de repositories publics
              example: 42
            followers:
              type: integer
              description: Nombre d'abonnés
              example: 150
            following:
              type: integer
              description: Nombre de personnes suivies
              example: 75
    ProjectRoleDto:
      type: object
      properties:
        roleTitle:
          type: string
          description: Titre du rôle recherché
          example: 'Frontend Developer'
        description:
          type: string
          description: Description détaillée du rôle
          example: "Développeur React expérimenté pour l'interface utilisateur"
        skillSet:
          type: array
          items:
            $ref: '#/components/schemas/TechStackDto'
          description: Compétences techniques requises pour ce rôle
        isFilled:
          type: boolean
          description: Indique si le rôle est déjà pourvu
          default: false
      required:
        - roleTitle
        - description
        - skillSet
        - isFilled
    ProjectMemberDto:
      type: object
      properties:
        userId:
          type: string
          description: ID de l'utilisateur membre du projet
          example: 'github_123456789'
        role:
          type: string
          description: Rôle de l'utilisateur dans le projet
          example: 'Frontend Developer'
        joinedAt:
          type: string
          format: date-time
          description: Date d'ajout au projet
          example: '2024-01-15T10:30:00.000Z'
        user:
          $ref: '#/components/schemas/GitHubUser'
      required:
        - userId
        - role
        - joinedAt
...
  securitySchemes:
    supertokens_session:
      type: apiKey
      in: cookie
      name: sAccessToken
      description: |
        Jeton de session SuperTokens pour l'authentification.
        Automatiquement géré par le navigateur via cookies HttpOnly.

        **Cookies utilisés :**
        - `sAccessToken` : Token d'accès principal
        - `sRefreshToken` : Token de rafraîchissement (optionnel selon la route)
    supertokens_refresh:
      type: apiKey
      in: cookie
      name: sRefreshToken
      description: |
        Jeton de rafraîchissement SuperTokens pour renouveler les sessions.
        Utilisé automatiquement par `/auth/session/refresh`.